Refactor storage file names to be more descriptive
